Updated with fix for Tabular RL. Still need implementations, but cleaned up interface and brought it inline with LinearRL.
