+++ src/JuliaRL.jl
@@ -8,10 +8,6 @@ include("FeatureCreators.jl")
 export Learning
 include("Learning.jl")
 
-# export LinearRL, TabularRL
-# include("LinearRL.jl")
-# include("TabularRL.jl")
-
 export ExperienceReplay, WeightedExperienceReplay, size, getindex, add!
 include("Replay.jl")
