+++ src/learning/Updates.jl
@@ -1,0 +1,18 @@
+
+mutable struct TD
+end
+
+mutable struct TDLambda
+end
+
+mutable struct VTrace
+end
+
+mutable struct Sarsa
+end
+
+mutable struct ExpectedSarsa
+end
+
+mutable struct WatkinsQ
+end
